+++ src/main/java/com/benmu/framework/BMWXApplication.java
@@ -30,7 +30,6 @@ public class BMWXApplication extends Application {
     private static BMWXApplication mInstance;
     private WXSDKInstance mMediator;
     private VersionChecker mVersionChecker;
-    public boolean isRecordHomeActivity;
 
     @Override
     public void onCreate() {